+++ R/tbl_likert.R
@@ -1,6 +1,8 @@
-#' Likert Summary
+#' Likert Summary Table
 #'
 #' \lifecycle{experimental}
+#' Creates a summary of Likert scales following the gtsummary structure
+#'
 #' @inheritParams gtsummary::tbl_summary
 #' @param statistic Formula or list of formulas specifying types of categorical
 #' statistics to display, see [gtsummary::tbl_summary()] help page,
@@ -9,8 +11,10 @@
 #' computed statistics, see [gtsummary::tbl_summary()] help page
 #' @param sort Sort table based on mean scores? Must be one of
 #' `c("default", "ascending", "descending")`
+#' @param x Table of class 'tbl_likert'
+#' @param ... not used
+#' @family gtsummary-related functions
 #' @export
-#' @name tbl_likert
 #' @examples
 #' df <-
 #'   tibble::tibble(
@@ -22,9 +26,15 @@
 #'       factor(levels = 1:3, labels = c("bad", "meh", "good")),
 #'   )
 #'
-#' tbl_likert(df) %>%
-#'   add_n() %>%
-#'   gtsummary::as_kable()
+#' tbl_likert_ex1 <-
+#'   tbl_likert(df) %>%
+#'   add_n()
+#' @export
+#' @name tbl_likert
+#' @section Example Output:
+#' \if{html}{Example 1}
+#'
+#' \if{html}{\figure{tbl_likert_ex1.png}{options: width=50\%}}
 NULL
 
 #' @export